🐛 fix(data-type) : changed data-type for the number

+++ src/com/aca/CheckMaxNumber.java
@@ -7,11 +7,11 @@ public class CheckMaxNumber {           //TASK#5
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter the first number");
-        int a = scanner.nextInt();
+        long a = scanner.nextLong();
         System.out.println("Enter the second number");
-        int b = scanner.nextInt();
+        long b = scanner.nextLong();
         System.out.println("Enter the third number");
-        int c = scanner.nextInt();
+        long c = scanner.nextLong();
 
         if (a > b && a > c) {
             System.out.println("Maximum number is  " + a);
